refactor: split marketing performance plane
